refactor(tplan): finalize task and presentation ownership (#172)
