+++ skills/tplan/resources/runtime-manifest.json
@@ -27,8 +27,11 @@
     "scripts/tplan_runtime.py",
     "scripts/tplan_errors.py",
     "scripts/tplan_identity.py",
+    "scripts/tplan_task_contract.py",
     "scripts/execution_time_metrics.py",
     "scripts/execution_cost_tree.py",
+    "scripts/execution_svg_renderer.py",
+    "scripts/execution_markdown_renderer.py",
     "scripts/codex_telemetry_activation.py",
     "scripts/codex_telemetry_adapter.py",
     "scripts/codex_telemetry_dispatcher.py",
@@ -41,8 +44,11 @@
     "scripts/tplan_runtime.py",
     "scripts/tplan_errors.py",
     "scripts/tplan_identity.py",
+    "scripts/tplan_task_contract.py",
     "scripts/execution_time_metrics.py",
     "scripts/execution_cost_tree.py",
+    "scripts/execution_svg_renderer.py",
+    "scripts/execution_markdown_renderer.py",
     "scripts/codex_telemetry_activation.py",
     "scripts/codex_telemetry_adapter.py",
     "scripts/codex_telemetry_dispatcher.py",
