+++ skills/tplan/scripts/execution_cost_tree.py
@@ -2451,1055 +2451,20 @@ def _compact_node_summary(node: dict[str, Any], reasons: set[str]) -> str:
 
 
 def render_compact_text(report: dict[str, Any]) -> str:
-    if report["view"] != "compact":
-        raise TplanError("text tree is only available for compact view")
-    mission = report["mission"]
-    mission_elapsed = _fmt_covered_duration(
-        mission["elapsed_ms"], mission["observed_elapsed_ms"], mission["elapsed_coverage"]
-    )
-    if mission["elapsed_coverage"] == "partial":
-        mission_elapsed = f"观测 {mission_elapsed}"
-    mission_token = _fmt_tokens_inline(mission["cost"])
-    mission_line = (
-        f"Mission · {_shorten(mission['title'], 60)} "
-        f"{_compact_status(mission['status'])} {mission_elapsed}"
-        f" · {_compact_cost_summary(mission['cost'])}"
-    )
-    if mission_token != "—":
-        mission_line += f" · Tok {mission_token}"
-    mission_line += f" · {short_attribution_label(mission['outcome_attribution'])}"
-    lines = [mission_line, "层级：[T] Task · [ST] SubTask · [P] Step"]
-    node_by_id = {node["id"]: node for node in report["nodes"]}
-    selection_reasons = {
-        task_id: set(reasons)
-        for task_id, reasons in report["trace"]["selection_reasons"].items()
-    }
-    order = {node["id"]: index for index, node in enumerate(report["nodes"])}
-    children: dict[str, list[str]] = defaultdict(list)
-    roots: list[str] = []
-    for node in report["nodes"]:
-        parent_id = node.get("parent_id")
-        if parent_id in node_by_id:
-            children[parent_id].append(node["id"])
-        else:
-            roots.append(node["id"])
-    for parent_id in children:
-        children[parent_id].sort(key=lambda task_id: order[task_id])
-    roots.sort(key=lambda task_id: order[task_id])
-
-    def visit(task_id: str, prefix: str, is_last: bool) -> None:
-        connector = "└─ " if is_last else "├─ "
-        lines.append(
-            prefix
-            + connector
-            + _compact_node_summary(node_by_id[task_id], selection_reasons.get(task_id, set()))
-        )
-        next_prefix = prefix + ("   " if is_last else "│  ")
-        child_ids = children.get(task_id, [])
-        for index, child_id in enumerate(child_ids):
-            visit(child_id, next_prefix, index == len(child_ids) - 1)
-
-    for index, task_id in enumerate(roots):
-        visit(task_id, "", index == len(roots) - 1)
-    lines.extend(
-        [
-            "",
-            (
-                f"显示 {report['trace']['visible_node_count']}/{report['trace']['total_node_count']}；"
-                f"省略 {report['trace']['hidden_node_count']}"
-            ),
-            (
-                f"{_compact_source_legend(mission['cost'])}；"
-                f"未精确记录 {_fmt_not_exactly_recorded(mission['elapsed_reconciliation'])}；"
-                "— 未采集 · ? 未知 · ≈ 估算 · ≥ 部分采集"
-            ),
-        ]
-    )
-    runtime_text = _runtime_diagnostic_text(report)
-    if runtime_text:
-        lines.extend(["", f"运行时告警：{runtime_text}"])
-    return "\n".join(lines) + "\n"
+    from execution_markdown_renderer import render_compact_text as _render
+    return _render(report)
 
 
 def render_svg(report: dict[str, Any]) -> str:
-    """Render a portrait SVG: chronological rows plus the real hierarchy overlay.
-
-    Vertical spacing follows event order, not elapsed duration. Every row therefore
-    remains readable even when the Mission contains long idle gaps. Exact relative
-    offsets are printed beside the axis, and each card has a linearly scaled range bar
-    against the same Mission duration.
-    """
-
-    if report["view"] == "compact":
-        raise TplanError("compact view uses a Unicode text tree; SVG is only available for standard/audit")
-    mission = report["mission"]
-    timeline = report["timeline"]
-    node_by_id = {node["id"]: node for node in report["nodes"]}
-    rows = timeline["rows"]
-    view = report["view"]
-    hotspot_by_id = {
-        item["task_id"]: item for item in report.get("duration_hotspots", {}).get("tasks", [])
-    }
-    time_coverage = timeline.get("offset_coverage") or report["trace"]["coverage"]
-    standard_node_lines = (
-        {
-            row["node_id"]: _standard_node_content_lines(
-                node_by_id[row["node_id"]],
-                row,
-                time_coverage=time_coverage,
-            )
-            for row in rows
-        }
-        if view == "standard"
-        else {}
-    )
-    standard_header_lines = (
-        _standard_mission_header_lines(report) if view == "standard" else []
-    )
-    width = 1180
-    header_x = 32
-    header_y = 24
-    header_width = width - 64
-    header_height = (
-        288
-        if view == "audit"
-        else max(168, 120 + 24 * len(standard_header_lines))
-    )
-    axis_x = 112
-    card_base_x = 270
-    depth_indent = 42
-    right_margin = 34
-    legend_y = header_y + header_height + 30
-    row_top = legend_y + 22
-    card_height = (
-        296
-        if view == "audit"
-        else max(
-            154,
-            54 + 25 * max(
-                (len(lines) for lines in standard_node_lines.values()),
-                default=4,
-            ),
-        )
-    )
-    row_gap_by_kind = {"task": 30, "subtask": 18, "step": 12}
-    row_ys: list[float] = []
-    row_cursor = float(row_top)
-    for index, row in enumerate(rows):
-        if index:
-            row_kind = node_by_id[row["node_id"]]["kind"]
-            row_cursor += row_gap_by_kind.get(row_kind, 16)
-        row_ys.append(row_cursor)
-        row_cursor += card_height
-    footer_height = 70
-    content_bottom = row_cursor if rows else row_top + card_height
-    height = content_bottom + footer_height
-    axis_start_y = row_ys[0] + card_height / 2 if rows else row_top + card_height / 2
-    axis_end_y = row_ys[-1] + card_height / 2 if rows else axis_start_y
-    mission_elapsed = timeline.get("window_elapsed_ms")
-    time_label = (
-        "实际相对时间"
-        if time_coverage == "exact"
-        else "已观测相对时间"
-        if time_coverage == "partial"
-        else "无执行时间刻度"
-    )
-    mission_status = STATUS_LABELS.get(mission["status"], mission["status"])
-    mission_cost = mission["cost"]
-    mission_reconciliation = mission["elapsed_reconciliation"]
-    report_title = _execution_presentation_title(report)
-    visible_node_label = "可见真实节点" if report["trace"]["projection"] else "真实节点"
-    coverage_warning_count = len(report["trace"].get("coverage_diagnostics", []))
-    runtime = report["runtime"]
-    time_domain_description = (
-        "左侧刻度显示 Mission 生命周期相对时间，时间条按同一 Mission 历时等比例显示"
-        if time_coverage == "exact"
-        else "左侧刻度只显示已观测窗口相对时间，时间条不代表完整 Mission 历时"
-        if time_coverage == "partial"
-        else "没有执行时间观测，卡片按声明拓扑顺序排列且不显示时间条"
-    )
-
-    lines = [
-        '<?xml version="1.0" encoding="UTF-8"?>',
-        (
-            f'<svg xmlns="http://www.w3.org/2000/svg" width="{width}" height="{height}" '
-            f'viewBox="0 0 {width} {height}" role="img" aria-labelledby="tplan-title tplan-desc" '
-            f'data-layout="vertical-execution-timeline" '
-            f'data-view="{_html(view)}" '
-            f'data-schema-version="{_html(report["schema_version"])}">'
-        ),
-        f'<title id="tplan-title">{_html(mission["title"])} · {_html(report_title)}</title>',
-        (
-            '<desc id="tplan-desc">纵向排列真实任务节点；'
-            f'{time_domain_description}；弱配色折线保留真实父子关系；'
-            '缩进、6/4/2px 主干与分支结构线、短层级标记、类型牌及标题字重共同区分 Task、SubTask 与 Step；'
-            'Task 主干与分支全线不透明以保持同色；父子线共享接头由父级等效色不透明覆盖，'
-            '避免子级颜色透出混色；'
-            '状态颜色集中在状态标签，异常节点才使用状态底色；中性耗时排名标签不表示失败。</desc>'
-        ),
-        "<style>",
-        "text{font-family:-apple-system,BlinkMacSystemFont,'Segoe UI','PingFang SC','Microsoft YaHei',sans-serif}",
-        ".report-title{font-size:22px;font-weight:700;fill:#ffffff}",
-        ".mission-title{font-size:18px;font-weight:650;fill:#dbeafe}",
-        ".mission-meta{font-size:14px;fill:#dbeafe}",
-        ".legend{font-size:13px;fill:#475569}",
-        ".time-label{font-size:12px;font-variant-numeric:tabular-nums;fill:#475569}",
-        ".node-title-task{font-size:16px;font-weight:700;fill:#0f172a}",
-        ".node-title-subtask{font-size:15px;font-weight:650;fill:#0f172a}",
-        ".node-title-step{font-size:14px;font-weight:600;fill:#0f172a}",
-        ".kind-label{font-size:11px;font-weight:750;letter-spacing:.35px}",
-        ".node-meta{font-size:12.5px;fill:#475569}",
-        ".node-metric{font-size:13px;fill:#1e293b}",
-        ".node-result{font-size:12.5px;fill:#334155}",
-        ".status-label{font-size:12px;font-weight:650}",
-        ".duration-hotspot-label{font-size:12px;font-weight:650;fill:#475569}",
-        ".tree-edge{fill:none;stroke-linecap:round;stroke-linejoin:round}",
-        ".time-guide{stroke:#cbd5e1;stroke-width:1;stroke-dasharray:3 5}",
-        ".axis{stroke:#64748b;stroke-width:2}",
-        ".range-track{fill:#e8edf3}",
-        "</style>",
-        f'<rect width="{width}" height="{height}" fill="#f8fafc"/>',
-        (
-            f'<rect x="{header_x}" y="{header_y}" width="{header_width}" height="{header_height}" '
-            'rx="16" fill="#172554" stroke="#3b82f6" stroke-width="1.5"/>'
-        ),
-        _svg_text(58, 58, report_title, "report-title"),
-        _svg_text(58, 88, _shorten(mission["title"], 72), "mission-title"),
-        *(
-            [
-                _svg_text(
-                    58,
-                    116 + index * 24,
-                    _shorten(part, 135),
-                    "mission-meta",
-                )
-                for index, part in enumerate(standard_header_lines)
-            ]
-            if view == "standard"
-            else [
-                _svg_text(
-                    58,
-                    116,
-                    (
-                        f"Mission · {mission_status} · "
-                        f"{_elapsed_scope_label(mission['elapsed_coverage'], mission=True)} "
-                        f"{_fmt_covered_duration(mission['elapsed_ms'], mission['observed_elapsed_ms'], mission['elapsed_coverage'])}"
-                    ),
-                    "mission-meta",
-                ),
-                _svg_text(
-                    58,
-                    140,
-                    (
-                        f"LLM调用累计 {_fmt_kind_duration(mission_cost, LLM_KINDS, host_label='调用端实测')}"
-                        f" · 脚本累计 {_fmt_kind_duration(mission_cost, SCRIPT_KINDS)}"
-                    ),
-                    "mission-meta",
-                ),
-                _svg_text(
-                    58,
-                    164,
-                    (
-                        f"工具累计 {_fmt_kind_duration(mission_cost, TOOL_KINDS)}"
-                        f" · 等待累计 {_fmt_kind_duration(mission_cost, WAIT_KINDS)}"
-                        f" · 未被精确记录 {_fmt_not_exactly_recorded(mission_reconciliation)}"
-                    ),
-                    "mission-meta",
-                ),
-                _svg_text(
-                    58,
-                    188,
-                    (
-                        f"Token {_fmt_tokens_compact(mission_cost)}"
-                        f" · {visible_node_label} {report['trace']['visible_node_count']}/{report['trace']['total_node_count']}"
-                        f" · 产出归因 P{len(mission['outcome_attribution']['countable_progress'])}"
-                        f"/C{len(mission['outcome_attribution']['constraint_deltas'])}"
-                        f" · 覆盖 {report['trace']['coverage']}"
-                        + (
-                            f" · 生命周期告警 {coverage_warning_count}"
-                            if coverage_warning_count
-                            else ""
-                        )
-                        + f" · 运行时 {runtime['status']}"
-                        + f" · {view}"
-                    ),
-                    "mission-meta",
-                ),
-                *[
-                    _svg_text(
-                        58,
-                        212 + index * 24,
-                        ("Mission 审计：" if index == 0 else "")
-                        + _shorten(part, 135),
-                        "mission-meta",
-                    )
-                    for index, part in enumerate(
-                        attribution_audit_lines(mission["outcome_attribution"])
-                    )
-                ],
-            ]
-        ),
-        _svg_text(
-            32,
-            legend_y,
-            (
-                f"纵向=首次执行；左侧是{time_label}；蓝灰条=起止/持续；"
-                "层级线=Task 主干 6 / SubTask 分支 4 / Step 末梢 2px；中性标签=Task 耗时排名。"
-            ),
-            "legend",
-        ),
-    ]
-
-    if rows:
-        lines.append(
-            f'<line x1="{axis_x}" y1="{axis_start_y}" x2="{axis_x}" y2="{axis_end_y}" class="axis"/>'
-        )
-
-    positions: dict[str, dict[str, float]] = {}
-    for index, row in enumerate(rows):
-        node = node_by_id[row["node_id"]]
-        card_x = card_base_x + row["depth"] * depth_indent
-        card_y = row_ys[index]
-        card_width = width - right_margin - card_x
-        center_y = card_y + card_height / 2
-        positions[node["id"]] = {
-            "x": card_x,
-            "y": card_y,
-            "width": card_width,
-            "height": card_height,
-            "center_y": center_y,
-            "kind": node["kind"],
-        }
-        lines.extend(
-            [
-                f'<line x1="{axis_x + 8}" y1="{center_y}" x2="{card_x - 10}" y2="{center_y}" class="time-guide"/>',
-                f'<circle cx="{axis_x}" cy="{center_y}" r="5" fill="#ffffff" stroke="#475569" stroke-width="2"/>',
-                _svg_text(
-                    axis_x - 12,
-                    center_y + 4,
-                    _fmt_timeline_offset(row["start_offset_ms"], time_coverage),
-                    "time-label",
-                    anchor="end",
-                ),
-            ]
-        )
-
-    # SVG uses painter's order. Paint thinner child branches first so the thicker
-    # parent edge owns their shared junction instead of being visibly cut by it.
-    edge_paint_order = {"step": 0, "subtask": 1, "task": 2}
-    edge_styles = {
-        "task": {
-            "width": 6.0,
-            "color": "#64748b",
-            "opacity": 1.0,
-            "opaque_equivalent": "#64748b",
-        },
-        "subtask": {
-            "width": 4.0,
-            "color": "#8b5cf6",
-            "opacity": 0.72,
-            "opaque_equivalent": "#aa88f8",
-        },
-        "step": {
-            "width": 2.0,
-            "color": "#60a5fa",
-            "opacity": 0.84,
-            "opaque_equivalent": "#78b3fa",
-        },
-    }
-    painted_edges = sorted(
-        report["tree_edges"],
-        key=lambda edge: edge_paint_order.get(
-            positions.get(edge["to"], {}).get("kind", ""), -1
-        ),
-    )
-    junction_caps: dict[tuple[str, float, float, float], dict[str, Any]] = {}
-    for edge in painted_edges:
-        child_position = positions.get(edge["to"])
-        if child_position is None:
-            continue
-        child_x = child_position["x"]
-        child_y = child_position["center_y"]
-        edge_style = edge_styles.get(
-            child_position["kind"],
-            {
-                "width": 2.0,
-                "color": "#94a3b8",
-                "opacity": 0.72,
-                "opaque_equivalent": "#a7b0bd",
-            },
-        )
-        if edge["from"] == "mission":
-            branch_x = card_base_x - 28
-            path = f"M {branch_x} {header_y + header_height} V {child_y} H {child_x}"
-        else:
-            parent_position = positions.get(edge["from"])
-            if parent_position is None:
-                continue
-            parent_x = parent_position["x"]
-            parent_y = parent_position["center_y"]
-            branch_x = min(parent_x, child_x) - 18
-            path = f"M {parent_x} {parent_y} H {branch_x} V {child_y} H {child_x}"
-            parent_kind = parent_position["kind"]
-            parent_style = edge_styles.get(parent_kind)
-            if parent_style is not None:
-                junction_caps[(edge["from"], parent_x, parent_y, branch_x)] = {
-                    "parent_id": edge["from"],
-                    "kind": parent_kind,
-                    "x1": branch_x,
-                    "x2": parent_x,
-                    "y": parent_y,
-                    "style": parent_style,
-                }
-        lines.append(
-            (
-                f'<path d="{path}" class="tree-edge" stroke="{edge_style["color"]}" '
-                f'stroke-width="{edge_style["width"]}" opacity="{edge_style["opacity"]}" '
-                f'data-tree-from="{_html(edge["from"])}" data-tree-to="{_html(edge["to"])}" '
-                f'data-child-kind="{_html(child_position["kind"])}"/>'
-            )
-        )
-
-    for cap in sorted(
-        junction_caps.values(),
-        key=lambda item: edge_paint_order.get(item["kind"], -1),
-    ):
-        cap_style = cap["style"]
-        lines.append(
-            (
-                f'<line x1="{cap["x1"]}" y1="{cap["y"]}" x2="{cap["x2"]}" y2="{cap["y"]}" '
-                f'class="tree-edge junction-cap" stroke="{cap_style["opaque_equivalent"]}" '
-                f'stroke-width="{cap_style["width"]}" opacity="1" '
-                f'data-junction-parent="{_html(cap["parent_id"])}" '
-                f'data-parent-kind="{_html(cap["kind"])}"/>'
-            )
-        )
-
-    for row in rows:
-        node = node_by_id[row["node_id"]]
-        hotspot = hotspot_by_id.get(node["id"])
-        position = positions[node["id"]]
-        card_x = position["x"]
-        card_y = position["y"]
-        card_width = position["width"]
-        status_fill, status_stroke, text_color = _svg_status_palette(node["status"])
-        if node["status"] == "completed":
-            card_fill = "#ffffff"
-            card_stroke = "#d8e0ea"
-            card_stroke_width = 1.2
-        else:
-            card_fill = status_fill
-            card_stroke = status_stroke
-            card_stroke_width = 1.5
-        cost, reconciliation, scope_label = _node_cost_view(node)
-        kind_label = {"task": "Task", "subtask": "SubTask", "step": "Step"}.get(
-            node["kind"], str(node["kind"] or "Node")
-        )
-        kind_style = {
-            "task": {
-                "label": "TASK",
-                "fill": "#475569",
-                "stroke": "#475569",
-                "text": "#ffffff",
-            },
-            "subtask": {
-                "label": "SUBTASK",
-                "fill": "#f3e8ff",
-                "stroke": "#8b5cf6",
-                "text": "#6d28d9",
-            },
-            "step": {
-                "label": "STEP",
-                "fill": "#eff6ff",
-                "stroke": "#3b82f6",
-                "text": "#1d4ed8",
-            },
-        }.get(
-            node["kind"],
-            {
-                "label": kind_label.upper(),
-                "fill": "#f8fafc",
-                "stroke": "#94a3b8",
-                "text": "#475569",
-            },
-        )
-        hierarchy_accent = {
-            "task": {"color": "#64748b", "width": 6},
-            "subtask": {"color": "#8b5cf6", "width": 4},
-            "step": {"color": "#60a5fa", "width": 2},
-        }.get(node["kind"], {"color": "#94a3b8", "width": 2})
-        title_class = {
-            "task": "node-title-task",
-            "subtask": "node-title-subtask",
-            "step": "node-title-step",
-        }.get(node["kind"], "node-title-subtask")
-        status = STATUS_LABELS.get(node["status"], node["status"])
-        icon = STATUS_ICONS.get(node["status"], "•")
-        title_limit = 32 if view != "audit" else 46
-        title = f"{_shorten(node['title'], title_limit)} · {node['id']}"
-        elapsed = _fmt_covered_duration(
-            node["elapsed_ms"], node["observed_elapsed_ms"], node["elapsed_coverage"]
-        )
-        range_text = (
-            f"{_fmt_timeline_offset(row['start_offset_ms'], time_coverage)} → "
-            f"{_fmt_timeline_offset(row['finish_offset_ms'], time_coverage)}"
-        )
-        hotspot_rank_attr = f' data-duration-hotspot-rank="{hotspot["rank"]}"' if hotspot else ""
-        lines.extend(
-            [
-                (
-                    f'<g id="node-{_html(node["id"])}" class="task-card" '
-                    f'data-task-id="{_html(node["id"])}" data-depth="{row["depth"]}" '
-                    f'data-start-offset-ms="{row["start_offset_ms"] if row["start_offset_ms"] is not None else ""}" '
-                    f'data-finish-offset-ms="{row["finish_offset_ms"] if row["finish_offset_ms"] is not None else ""}"'
-                    f'{hotspot_rank_attr}>'
-                ),
-                (
-                    f'<rect x="{card_x}" y="{card_y}" width="{card_width}" height="{card_height}" '
-                    f'rx="10" fill="{card_fill}" stroke="{card_stroke}" stroke-width="{card_stroke_width}"/>'
-                ),
-                (
-                    f'<rect x="{card_x + 3}" y="{card_y + 13}" width="{hierarchy_accent["width"]}" '
-                    f'height="28" rx="{hierarchy_accent["width"] / 2}" fill="{hierarchy_accent["color"]}" '
-                    f'class="hierarchy-accent" data-node-kind="{_html(node["kind"])}"/>'
-                ),
-                (
-                    f'<rect x="{card_x + 18}" y="{card_y + 13}" width="72" height="24" rx="6" '
-                    f'fill="{kind_style["fill"]}" stroke="{kind_style["stroke"]}" stroke-width="1.25"/>'
-                ),
-                (
-                    f'<text x="{card_x + 54}" y="{card_y + 30}" class="kind-label" text-anchor="middle" '
-                    f'fill="{kind_style["text"]}">{_html(kind_style["label"])}</text>'
-                ),
-                _svg_text(card_x + 104, card_y + 29, title, title_class),
-                (
-                    f'<rect x="{card_x + card_width - 94}" y="{card_y + 13}" width="76" height="24" '
-                    f'rx="12" fill="{status_stroke}" opacity=".13"/>'
-                ),
-                (
-                    f'<text x="{card_x + card_width - 56}" y="{card_y + 30}" class="status-label" '
-                    f'text-anchor="middle" fill="{text_color}">{_html(icon + " " + status)}</text>'
-                ),
-            ]
-        )
-        if hotspot:
-            hotspot_badge_x = card_x + card_width - 224
-            lines.extend(
-                [
-                    (
-                        f'<rect x="{hotspot_badge_x}" y="{card_y + 13}" width="118" height="24" '
-                        'rx="12" fill="#f8fafc" stroke="#94a3b8" stroke-width="1.25"/>'
-                    ),
-                    (
-                        f'<text x="{hotspot_badge_x + 59}" y="{card_y + 30}" '
-                        'class="duration-hotspot-label" text-anchor="middle">'
-                        f'耗时排名 #{hotspot["rank"]}</text>'
-                    ),
-                ]
-            )
-        if view == "standard":
-            lines.extend(
-                _svg_text(
-                    card_x + 18,
-                    card_y + 54 + index * 25,
-                    text,
-                    class_name,
-                )
-                for index, (text, class_name) in enumerate(
-                    standard_node_lines[node["id"]]
-                )
-            )
-            range_y = card_y + card_height - 14
-        else:
-            time_metric = _svg_text(
-                card_x + 18,
-                card_y + 79,
-                (
-                    f"时间 {range_text} · "
-                    f"{_elapsed_scope_label(node['elapsed_coverage'])} {elapsed}"
-                    f" · 未被精确记录 {_fmt_not_exactly_recorded(reconciliation)}"
-                ),
-                "node-metric",
-            )
-            lines.extend(
-                [
-                    _svg_text(
-                        card_x + 18,
-                        card_y + 54,
-                        _svg_node_status_details(node, scope_label),
-                        "node-meta",
-                    ),
-                    time_metric,
-                    _svg_text(
-                        card_x + 18,
-                        card_y + 104,
-                        (
-                            f"LLM调用累计 {_fmt_kind_duration(cost, LLM_KINDS, host_label='调用端实测')}"
-                            f" · 脚本累计 {_fmt_kind_duration(cost, SCRIPT_KINDS)}"
-                        ),
-                        "node-metric",
-                    ),
-                    _svg_text(
-                        card_x + 18,
-                        card_y + 129,
-                        (
-                            f"工具累计 {_fmt_kind_duration(cost, TOOL_KINDS)}"
-                            f" · 等待累计 {_fmt_kind_duration(cost, WAIT_KINDS)}"
-                            f" · Token {_fmt_tokens_compact(cost)}"
-                        ),
-                        "node-metric",
-                    ),
-                    _svg_text(
-                        card_x + 18,
-                        card_y + 151,
-                        f"结果：{_shorten(node['outcome_summary'], 82) if node['outcome_summary'] else '未记录'}",
-                        "node-result",
-                    ),
-                    _svg_text(
-                        card_x + 18,
-                        card_y + 174,
-                        f"产出归因：{_shorten(attribution_text(node['outcome_attribution']), 74)}",
-                        "node-result",
-                    ),
-                ]
-            )
-            lines.extend(
-                _svg_text(
-                    card_x + 18,
-                    card_y + 197 + index * 23,
-                    ("审计：" if index == 0 else "") + _shorten(part, 105),
-                    "node-meta",
-                )
-                for index, part in enumerate(
-                    attribution_audit_lines(node["outcome_attribution"])
-                )
-            )
-            range_y = card_y + 282
-
-        track_x = card_x + 18
-        track_width = card_width - 36
-        lines.append(
-            f'<rect x="{track_x}" y="{range_y}" width="{track_width}" height="6" rx="3" class="range-track"/>'
-        )
-        start_offset = row["start_offset_ms"]
-        finish_offset = row["finish_offset_ms"]
-        if mission_elapsed and start_offset is not None and finish_offset is not None:
-            start_fraction = min(1.0, max(0.0, start_offset / mission_elapsed))
-            finish_fraction = min(1.0, max(start_fraction, finish_offset / mission_elapsed))
-            range_x = min(track_x + track_width - 4.0, track_x + start_fraction * track_width)
-            range_width = max(4.0, (finish_fraction - start_fraction) * track_width)
-            range_width = min(range_width, track_x + track_width - range_x)
-            range_y_rendered = range_y
-            range_height = 6
-            range_fill = "#64748b"
-            range_class = "node-range"
-            lines.append(
-                (
-                    f'<rect x="{range_x:.2f}" y="{range_y_rendered}" width="{range_width:.2f}" '
-                    f'height="{range_height}" rx="4" fill="{range_fill}" class="{range_class}" '
-                    f'data-task-id="{_html(node["id"])}"/>'
-                )
-            )
-        lines.append("</g>")
-
-    footer_y = content_bottom + 40
-    if time_coverage == "exact":
-        footer_text = (
-            f"Mission 结束 {_fmt_timeline_offset(mission_elapsed, time_coverage)}"
-            " · 纵向行距不代表持续时间；"
-            "精确时间由刻度、节点起止值及统一比例时间条表达。"
-        )
-    elif time_coverage == "partial":
-        footer_text = (
-            f"观测窗口结束 {_fmt_timeline_offset(mission_elapsed, time_coverage)}"
-            " · 纵向行距不代表持续时间；"
-            "刻度和时间条仅表达已观测事件的相对位置，不代表完整 Mission 历时。"
-        )
-    else:
-        footer_text = (
-            "没有执行时间观测；纵向行距不代表持续时间；"
-            "卡片仅保留声明拓扑、状态和异常信号，未显示成本不按零处理。"
-        )
-    lines.extend(
-        [
-            f'<line x1="32" y1="{footer_y - 24}" x2="{width - 32}" y2="{footer_y - 24}" stroke="#e2e8f0"/>',
-            _svg_text(
-                32,
-                footer_y,
-                footer_text,
-                "legend",
-            ),
-        ]
-    )
-    if report["trace"]["hidden_node_count"]:
-        lines.append(
-            _svg_text(
-                32,
-                footer_y + 24,
-                f"投影视图省略 {report['trace']['hidden_node_count']} 个真实节点；Standard/Audit 可查看完整拓扑。",
-                "legend",
-            )
-        )
-    lines.append("</svg>")
-    return "\n".join(lines) + "\n"
-
-
-def _markdown_cell(value: Any) -> str:
-    return str(value).replace("|", "\\|").replace("\n", " ")
-
-
-def _coverage_note(report: dict[str, Any]) -> str:
-    coverage = report["trace"]["coverage"]
-    if coverage == "exact":
-        return (
-            "生命周期追踪可从 Mission 初始化完整回放到当前快照；"
-            "各类成本仍只包含宿主实际上报的 span。"
-        )
-    if coverage == "partial":
-        return (
-            "生命周期追踪不完整或与当前快照不一致；这里只显示已观测窗口，"
-            "不把最后一条事件当作 Mission 完成时间。"
-        )
-    return "没有执行轨迹；当前仅展示 Mission 快照，时间与 Token 成本保持未知。"
-
-
-def _coverage_diagnostic_text(report: dict[str, Any]) -> str | None:
-    diagnostics = report["trace"].get("coverage_diagnostics", [])
-    if not diagnostics:
-        return None
-    return "；".join(
-        f"{item['code']}: {item['message']}" for item in diagnostics
-    )
-
-
-def _runtime_diagnostic_text(report: dict[str, Any]) -> str | None:
-    diagnostics = report["runtime"].get("diagnostics", [])
-    if not diagnostics:
-        return None
-    return "；".join(
-        f"{item['code']}: {item['message']}" for item in diagnostics
-    )
-
-
-def _append_telemetry_capture_markdown(
-    lines: list[str], report: dict[str, Any]
-) -> None:
-    capture = report["telemetry_capture"]
-    activation = capture["activation"]
-    labels = {
-        "local_tools": "本地工具/脚本",
-        "hosted_tools": "托管工具",
-        "model_turns": "模型/Turn",
-        "tokens": "Token",
-        "waits": "等待",
-        "subagents": "SubAgent",
-    }
-    lines.extend(
-        [
-            "",
-            "## Codex 遥测覆盖",
-            "",
-            (
-                f"绑定：`{capture['binding']['status']}`；范围："
-                f"`{capture['binding'].get('scope') or 'none'}`；generation："
-                f"`{capture['binding'].get('generation') or 'none'}`。"
-                "没有观测值的类别保持 `not_reported`，不会按零处理。"
-            ),
-            "",
-            (
-                f"激活：`{activation['status']}`；当前 surface："
-                f"`{activation.get('active_surface') or 'none'}`；"
-                f"原因：{activation['reason']}。"
-            ),
-            "",
-            "| Surface | 激活状态 | Codex build/version | Hook source | 原因 |",
-            "| --- | --- | --- | --- | --- |",
-        ]
-    )
-    for surface_name, surface in activation["surfaces"].items():
-        source = surface.get("source")
-        source_text = "none"
-        if isinstance(source, dict):
-            source_text = (
-                f"{source['scope']}:{source['path']}；"
-                f"hash={source.get('sha256') or 'absent'}；"
-                f"enumerated={str(source['enumerated']).lower()}；"
-                f"trust={','.join(source['trust_statuses']) or 'none'}；"
-                f"enabled={source.get('enabled')}"
-            )
-        build = surface.get("host_build") or surface.get("codex_version") or "none"
-        lines.append(
-            "| "
-            + " | ".join(
-                (
-                    _markdown_cell(surface_name),
-                    _markdown_cell(surface["status"]),
-                    _markdown_cell(build),
-                    _markdown_cell(source_text),
-                    _markdown_cell(surface["reason"]),
-                )
-            )
-            + " |"
-        )
-    lines.extend(
-        [
-            "",
-            "| 通道 | 状态 | 已完成 span | 原因 |",
-            "| --- | --- | ---: | --- |",
-        ]
-    )
-    for name, channel in capture["channels"].items():
-        lines.append(
-            "| "
-            + " | ".join(
-                (
-                    labels[name],
-                    _markdown_cell(channel["status"]),
-                    str(channel["observed_span_count"]),
-                    _markdown_cell(channel["reason"]),
-                )
-            )
-            + " |"
-        )
+    from execution_svg_renderer import render_svg as _render
+    return _render(report)
 
 
 def render_markdown(report: dict[str, Any], *, timeline_svg_ref: str | None = None) -> str:
-    if report["view"] == "compact":
-        if timeline_svg_ref is not None:
-            raise TplanError("compact Markdown is a Unicode text tree and has no SVG sidecar")
-        lines = [
-            "# TPlan 执行摘要",
-            "",
-            f"> {_coverage_note(report)}",
-            "",
-            "```text",
-            render_compact_text(report).rstrip(),
-            "```",
-            "",
-            "Task/SubTask 成本为其真实子树累计；Step 为直接成本。没有出现的下级节点只是被省略，没有被合并。",
-            "",
-            "LLM、脚本、工具和等待是累计资源时间，可能相互重叠；逐节点详情见 Standard/Audit。",
-        ]
-        diagnostic_text = _coverage_diagnostic_text(report)
-        if diagnostic_text:
-            lines[3:3] = [f"> 覆盖告警：{diagnostic_text}", ""]
-        runtime_text = _runtime_diagnostic_text(report)
-        if runtime_text:
-            lines[3:3] = [f"> 运行时告警：{runtime_text}", ""]
-        return "\n".join(lines).rstrip() + "\n"
-
-    mission = report["mission"]
-    cost = mission["cost"]
-    overhead = report["overhead"]["cost"]
-    presentation_title = _execution_presentation_title(report)
-    markdown_title = {
-        "exact": "# TPlan 实际执行与成本树",
-        "partial": "# TPlan 已观测执行窗口与成本树",
-        "snapshot_only": "# TPlan Mission 结构快照与成本边界",
-    }[report["trace"]["coverage"]]
-    lines = [
-        markdown_title,
-        "",
-        f"> {_coverage_note(report)}",
-        "",
-    ]
-    diagnostic_text = _coverage_diagnostic_text(report)
-    if diagnostic_text:
-        lines.extend([f"> 覆盖告警：{diagnostic_text}", ""])
-    runtime_text = _runtime_diagnostic_text(report)
-    if runtime_text:
-        lines.extend([f"> 运行时告警：{runtime_text}", ""])
-    if timeline_svg_ref:
-        lines.extend(
-            [
-                f"![{presentation_title}](<{timeline_svg_ref}>)",
-                "",
-            ]
-        )
-    else:
-        inline_svg = render_svg(report).split("\n", 1)[1].rstrip()
-        lines.extend([inline_svg, ""])
-    if report["view"] == "audit":
-        _append_telemetry_capture_markdown(lines, report)
-    lines.append("")
-    visible_node_label = "可见真实节点" if report["trace"]["projection"] else "真实节点"
-    lines.append(
-        f"视图：`{report['view']}`；布局：`vertical_execution_timeline`；"
-        f"{visible_node_label}：{report['trace']['visible_node_count']}/{report['trace']['total_node_count']}。"
-    )
-    if report["trace"]["hidden_node_count"]:
-        lines.append(
-            f"这是投影视图，省略了 {report['trace']['hidden_node_count']} 个真实节点；"
-            "使用 `--view standard` 或 `--view audit` 查看完整拓扑。"
-        )
-    mission_attribution = mission["outcome_attribution"]
-    node_yield_counts = mission_attribution.get("node_yield_counts", {})
-    writeback_only_nodes = node_yield_counts.get("writeback_only", 0)
-    telemetry_only_nodes = node_yield_counts.get("telemetry_only", 0)
-    lines.append(
-        "产出归因："
-        f"{len(mission_attribution['countable_progress'])} 项可计推进 · "
-        f"{len(mission_attribution['constraint_deltas'])} 项关键约束 · "
-        f"{writeback_only_nodes} 个仅状态写回节点 · {telemetry_only_nodes} 个仅遥测节点。"
-    )
-    lines.extend(
-        [
-            "",
-            "口径：覆盖 exact 时，实际历时按可回放生命周期的开始到结束计算；覆盖 partial 时，"
-            "这里只显示已观测窗口，不把它当作 Mission 完成时间。LLM 调用、脚本、工具和等待显示的是"
-            "各自累计资源时间，嵌套或并行时不可直接相加。调用端实测覆盖完整模型请求，可能包含"
-            "排队、网络和流式传输，不等于平台内部纯推理时间。未被精确记录 = 实际历时减去已完成且"
-            "时间来源精确的区间并集；它不自动属于 LLM、脚本或其他类别。已缓存输入包含在输入 Token 中，"
-            "不会重复累计。",
-        ]
-    )
-    if report["trace"]["open_span_count"]:
-        lines.append(
-            f"未结束调用：{report['trace']['open_span_count']} 个；已有入口记录但没有配对结束记录，"
-            "因此不计入累计成本。"
-        )
-    if report["view"] == "audit" and cost["envelope_span_count"]:
-        lines.append(
-            f"Agent turn 包络：{cost['envelope_span_count']} 个 span，"
-            f"{_fmt_kind_duration(cost, {'agent_turn'})}；仅用于审计，不计入 LLM 调用累计或可加资源时间。"
-        )
-    if report["view"] == "audit" and overhead["span_count"]:
-        lines.append(
-            f"Mission 级共享/未归属 span：{_fmt_duration(overhead['additive_resource_time_ms'])} 可加资源时间，"
-            f"{_fmt_tokens(overhead)} Token；这些成本不分摊到任务节点。"
-        )
-
-    notable = [
-        node
-        for node in report["nodes"]
-        if node["outcome_summary"]
-        or node["status"] in ABNORMAL_TASK_STATUSES
-        or node["attempts"] > 1
-        or node["direct_cost"]["error_span_count"]
-        or node["direct_open_span_count"]
-    ]
-    if notable and report["view"] == "audit":
-        lines.extend(["", "## 结果与异常", ""])
-        for node in notable:
-            details: list[str] = []
-            if node["outcome_summary"]:
-                details.append(_shorten(node["outcome_summary"], 180))
-            if node["attempts"] > 1:
-                details.append(f"执行 {node['attempts']} 次")
-            if node["direct_cost"]["error_span_count"]:
-                details.append(f"{node['direct_cost']['error_span_count']} 个错误 span")
-            if node["direct_open_span_count"]:
-                details.append(f"{node['direct_open_span_count']} 个未结束调用")
-            if not details:
-                details.append(STATUS_LABELS.get(node["status"], node["status"]))
-            lines.append(f"- {node['title']} (`{node['id']}`): {'; '.join(details)}")
-
-    if report["view"] == "audit":
-        mission_audit_relevant = any(
-            mission_attribution[field]
-            for field in (
-                "countable_progress",
-                "constraint_deltas",
-                "state_writebacks",
-                "unclassified_writebacks",
-                "warnings",
-            )
-        )
-        audit_nodes = [
-            node
-            for node in report["nodes"]
-            if node["outcome_attribution"]["unclassified_writebacks"]
-            or node["outcome_attribution"]["warnings"]
-            or node["outcome_attribution"]["countable_progress"]
-            or node["outcome_attribution"]["constraint_deltas"]
-        ]
-        if mission_audit_relevant or audit_nodes:
-            lines.extend(["", "## 产出归因审计", ""])
-            if mission_audit_relevant:
-                lines.append(
-                    f"- Mission (`{mission['id']}`): {attribution_text(mission_attribution)}；"
-                    f"{attribution_audit_text(mission_attribution)}。"
-                )
-            for node in audit_nodes:
-                attribution = node["outcome_attribution"]
-                evidence_ids = sorted(
-                    {
-                        evidence_id
-                        for bucket in (
-                            "countable_progress",
-                            "constraint_deltas",
-                            "unclassified_writebacks",
-                        )
-                        for item in attribution[bucket]
-                        for evidence_id in item.get("evidence_ids", [])
-                    }
-                )
-                commit_ids = sorted(
-                    {
-                        commit_id
-                        for bucket in (
-                            "countable_progress",
-                            "constraint_deltas",
-                            "state_writebacks",
-                            "unclassified_writebacks",
-                        )
-                        for item in attribution[bucket]
-                        for commit_id in item.get("commit_ids", [])
-                    }
-                )
-                reasons = [
-                    item.get("reason")
-                    for item in attribution["unclassified_writebacks"]
-                    if isinstance(item.get("reason"), str)
-                ]
-                warning_codes = [item["code"] for item in attribution["warnings"]]
-                lines.append(
-                    f"- {node['title']} (`{node['id']}`): {attribution_text(attribution)}；"
-                    f"evidence={evidence_ids or ['none']}；commit={commit_ids or ['none']}；"
-                    f"unclassified={reasons or ['none']}；warnings={warning_codes or ['none']}。"
-                )
-        lines.extend(
-            [
-                "",
-                "## 审计明细",
-                "",
-                (
-                    "| 顺序 | 节点 | 状态 | 产出归因 | "
-                    f"{_elapsed_scope_label(report['trace']['coverage'])} | 活跃时间 | 次数 | "
-                    "直接资源 | 子树资源 | 子树未精确记录 | 证据 | 产物 |"
-                ),
-                "| ---: | --- | --- | --- | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: |",
-            ]
-        )
-        for node in report["nodes"]:
-            lines.append(
-                "| "
-                + " | ".join(
-                    [
-                        str(node["execution_order"] or "—"),
-                        _markdown_cell(f"{node['title']} ({node['id']})"),
-                        _markdown_cell(node["actual_state"]),
-                        _markdown_cell(attribution_text(node["outcome_attribution"])),
-                        _fmt_covered_duration(
-                            node["elapsed_ms"], node["observed_elapsed_ms"], node["elapsed_coverage"]
-                        ),
-                        _fmt_covered_duration(
-                            node["active_duration_ms"],
-                            node["observed_active_duration_ms"],
-                            node["active_duration_source"],
-                        ),
-                        str(node["attempts"]),
-                        _fmt_resource_duration(node["direct_cost"]),
-                        _fmt_resource_duration(node["inclusive_cost"]),
-                        _fmt_not_exactly_recorded(node["subtree_elapsed_reconciliation"]),
-                        str(len(node["evidence_refs"])),
-                        str(len(node["artifact_refs"])),
-                    ]
-                )
-                + " |"
-            )
-    return "\n".join(lines).rstrip() + "\n"
+    from execution_markdown_renderer import render_markdown as _render
+    return _render(report, timeline_svg_ref=timeline_svg_ref)
 
 
 def render_json(report: dict[str, Any]) -> str:
-    return json.dumps(report, ensure_ascii=False, indent=2) + "\n"
+    from execution_markdown_renderer import render_json as _render
+    return _render(report)
